Fix handling of paragraphs in children_html
When converting bbcode tags to markdown, we run into trouble if there's new paragraphs (double newlines) in their content (in their `children_html`), this may not apply to all bbcode tags, but many of them e.g the basic [b] tag. We cannot output markdown looking like this:

```
**paragraph 1

paragraph 2**
```

It doesn't work (at least not in most markdown processors)

An ideal solution might be to output
```
**paragraph 1**

**parapgraph 2**
```
To do this properly will entail splitting the processing in two and re-applying our conversion logic to each part (and re-applying any other wrapping bbcode tags). There are data structures which might allow such tricks, but this is too complicated for me to take on right now.

The ugly workaround here is to swap in a `<br>` tag to give the same vertical spacing as a new paragraph while actually being part of the same paragraph.
```
**paragraph 1
<br>paragraph 2**
```
Ugly to look at, and ugly because we're outputting html when we're supposed to be converting to markdown. Furthermore this will only work where <br> tag is allowed (github and discourse at least). Better solutions welcome!