Allow :newlines => :to_br tag config
Allow a bbcode tag config to force all newlines into <br> html tags within its children_html. This is useful for outputting markdown headings as follows. Headings need to be all on one line even if the source bbcode allows newline chars.

`## My heading<br>continued on the next line.`

Outputting headings? This not part of the bbcode.org standard, but something we would like to be able to do as a custom "additional tag".