Add support for [code]
Add support for the [code] bbcode, making use of the new plain_text config option.

The code block is converted to be wrapped in "```" (on its own line), and crucially the block contents are unaltered by any further parsing.

TODO: In fact it will still convert ">" to "&gt;" and other html entities, which ideally it wouldn't, but when calling with `bbcode_to_md(false)` this is disactivated.

TODO: We can pass a language to achieve syntax highlighting, such as [code=javascript] in some bbcode flavours, and in the bbcode.org reference. This could be converted to "```javascript" in some markdown flavours. Currently it will just ignore the language information and convert a plain "```" block.