Add triple nesting test showing wrong output
Two test changes to begin tackling two different problems with nested quotes.

Firstly add a comment against the first test. This output does not work within discourse. Why talk about discourse specifically? The tag definition for [quote] doesn't do a proper "to markdown" conversion. It remains as bbcode, but with some newlines added. This is presumably to cater for the primary discourse use case, maybe once upon a time it did, but testing in discourse at the moment you cannot place the following test:

[quote=Kitten]
[quote=creatiu]f1[/quote]

f2[/quote]

We need a newline after 'f2' otherwise the oute quote is not rendered.

Secondly add a test for triple nesting of quotes. This test documents the current output we get, which is clearly not correct. There's no way the letters "EF" should end up next-to eachother. There's a problem with stacked handling of nesting.