Fix quotes output to have newlines which work in discourse
Fix the conversion of [quotes] to insert more newlines so that the output works in discourse (as it was presumably originally intended to).

Note: This changes all the output even for simple quote example, which may be surprising. The original output on one line works OK, but spacing them out onto multiple lines also looks OK. The problem we're fixing here only shows up in more complex nested cases.

Specifically this works when we paste it into descourse:

[quote=Kitten]

[quote=creatiu]
f1
[/quote]
f2
[/quote]

...but this does not render fully:

[quote=Kitten]
[quote=creatiu]f1[/quote]

f2[/quote]